+++ src/main/java/roomescape/payment/TossPaymentClient.java
@@ -1,11 +1,8 @@
 package roomescape.payment;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientResponseException;
-import roomescape.global.exception.IllegalRequestException;
 import roomescape.global.exception.PaymentFailException;
 import roomescape.payment.dto.request.PaymentConfirmRequest;
 import roomescape.payment.dto.response.PaymentConfirmResponse;
@@ -24,7 +21,6 @@ public class TossPaymentClient {
     private String password;
 
     private final RestClient restClient;
-    private final Logger log = LoggerFactory.getLogger(getClass());
 
     public TossPaymentClient(final RestClient restClient) {
         this.restClient = restClient;
@@ -39,14 +35,7 @@ public class TossPaymentClient {
                     .body(PaymentConfirmResponse.class);
         } catch (RestClientResponseException e) {
             TossErrorResponse errorResponse = e.getResponseBodyAs(TossErrorResponse.class);
-            if (errorResponse == null) {
-                throw new PaymentFailException("결제 승인 요청 중 오류가 발생했습니다.", e);
-            }
-            if (errorResponse.isClientError()) {
-                log.error("Toss Payment Client Error 발생: {}", errorResponse.message(), e);
-                throw new IllegalRequestException(errorResponse.message(), e);
-            }
-            throw new PaymentFailException(errorResponse.message(), e);
+            throw new PaymentFailException(errorResponse.code(), errorResponse.message(), e);
         } catch (Exception e) {
             throw new PaymentFailException("결제 승인 요청 중 오류가 발생했습니다.", e);
         }
